// src/types/models.ts
export interface ICrEvent {
    collateralReservationId: string;
    paymentAddress?: string;
    paymentAmount?: string;
    paymentReference?: string;
}

export interface IMintStatus {
    status: boolean;
    step: number;
}
// ...
export interface IMaxLots {
    maxLots: string;
    lotsLimited: boolean;
}
// ...
export interface IEstimateFee {
    estimatedFee: string;
    extraBTC: string;
}

export interface IRedemptionDefault {
    incomplete: boolean;
    remainingLots?: string;
}

export interface IUtxo {
    address?: string;
    hexTx: string;
    index: number;
    txid: number;
    utxoAddress: string;
    value: string;
    vout: number;
}

export interface IUtxoForTransaction {
    selectedUtxos: IUtxo[];
    returnAddresses: string[];
    estimatedFee: number;
}

export interface ILifetimeClaimed {
    fasset: string;
    claimed: string;
}

export interface IEcoSystemInfoSupplyByFasset {
    allLots: number;
    availableToMintLots: number;
    availableToMintUSD: string;
    fasset: string;
    minted: string;
    mintedPercentage: string;
    supply: string;
}

export interface IEcosystemInfo {
    agentCollateral: string;
    tvl: string;
    tvlPoolsNat: string;
    numTransactions: string;
    agentsInLiquidation: number;
    numAgents: number;
    numLiquidations: number;
    overCollaterazied: string;
    rewardsAvailableUSD: string;
    supplyByCollateral: {
        symbol: string;
        supply: string;
        supplyUSD: string;
    }[];
    supplyByFasset: IEcoSystemInfoSupplyByFasset[];
    totalCollateral: string;
    totalMinted: string;
    numMints: number;
    totalPoolRewardsPaidUSD: string;
    poolRewards: {
        fasset: string;
        rewards: string;
        rewardsUSD: string;
    }[];
    numHolders: string;
    numRedeems: number;
}

export interface ISubmitTx {
    hash: string;
}
// ...
export interface ICrStatus {
    status: boolean;
    accepted?: boolean;
    collateralReservationId: string;
    paymentAddress: string;
    paymentAmount: string;
    paymentReference: string;
}

export interface IReturnAddress {
    addresses: string[];
    estimatedFee: number;
    utxos: IUtxo[];
}

export interface IPrepareUtxo {
    psbt: string;
    selectedUtxos: IUtxo[];
}